Code refactored
1. Using camel case for file names
2. Ensuring best practices
