+++ lib/auth/firebase_auth.dart
