+++ lib/user_screens/admin/admin_dashboard.dart
@@ -2,7 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:weavershub/DashBoard/admin/accept-vendors.dart';
+import 'package:weavershub/user_screens/admin/accept_vendors.dart';
 
 class AdminDashBoard extends StatefulWidget {
   const AdminDashBoard({super.key});
