+++ pubspec.yaml
@@ -50,6 +50,7 @@ dependencies:
   intl:
   firebase_messaging: ^14.6.2
   flutter_local_notifications: ^14.1.1
+  cloud_firestore: ^4.17.5
 
 dev_dependencies:
   flutter_test:
